moved all activities to ui package
