+++ app/src/main/java/com/remote/universalirremote/ui/AcRemote.java
@@ -17,11 +17,13 @@
 //
 //
 
-package com.remote.universalirremote;
+package com.remote.universalirremote.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.remote.universalirremote.R;
 
 public class AcRemote extends AppCompatActivity {
