+++ app/src/main/java/com/remote/universalirremote/ui/NewDevice.java
@@ -18,7 +18,7 @@
 //
 
 
-package com.remote.universalirremote;
+package com.remote.universalirremote.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.NavUtils;
@@ -34,6 +34,8 @@ import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.Toast;
 
+import com.remote.universalirremote.Constant;
+import com.remote.universalirremote.R;
 import com.remote.universalirremote.database.DeviceData;
 import com.remote.universalirremote.database.DeviceInfoRepository;
