+++ app/src/main/res/layout/activity_ac_remote.xml
@@ -22,6 +22,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AcRemote">
+    tools:context=".ui.AcRemote">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
