+++ app/src/main/res/layout/activity_add_remote.xml
@@ -24,7 +24,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/back"
-    tools:context=".NewDevice">
+    tools:context=".ui.NewDevice">
 
     <TextView
         android:id="@+id/textViewDeviceSelect"
