+++ app/src/main/res/layout/activity_device_select.xml
@@ -23,7 +23,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/back"
-    tools:context=".DeviceSelect">
+    tools:context=".ui.DeviceSelect">
 
     <Spinner
         android:id="@+id/spnr_DeviceSelection"
